+++ src/commons/mod.rs
@@ -1,0 +1,8 @@
+//! This module provides common datastructures or CRDs shared between all the operators
+
+pub mod authentication;
+pub mod ldap;
+pub mod opa;
+pub mod resources;
+pub mod secret_class;
+pub mod tls;