+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod builder;
 pub mod cli;
 pub mod client;
+pub mod commons;
 pub mod config;
 pub mod crd;
 pub mod error;
@@ -8,7 +9,6 @@ pub mod label_selector;
 pub mod labels;
 pub mod logging;
 pub mod namespace;
-pub mod opa;
 pub mod pod_utils;
 pub mod product_config_utils;
 pub mod role_utils;