+++ CHANGELOG.md
@@ -4,17 +4,35 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.17.0] - 2022-04-14
+
+### Changed
+
+- product-config 0.3.1 -> 0.4.0 ([#373])
+- kube 0.70.0 -> 0.71.0 ([#372])
+
+[#372]: https://github.com/stackabletech/operator-rs/pull/372
+[#373]: https://github.com/stackabletech/operator-rs/pull/373
+
+## [0.16.0] - 2022-04-11
+
 ### Added
 
 - Export logs to Jaeger ([#360]).
+- Added common datastructures shared between all operators like `Tls` oder `AuthenticationClass` ([#366]).
+- Added helpers for env variables from Secrets or ConfigMaps ([#370]).
 
 ### Changed
 
 - BREAKING: `initialize_logging` now takes an app name and tracing target ([#360]).
+- BREAKING: Move opa struct to commons ([#369]).
 
 [#360]: https://github.com/stackabletech/operator-rs/pull/360
+[#366]: https://github.com/stackabletech/operator-rs/pull/366
+[#369]: https://github.com/stackabletech/operator-rs/pull/369
+[#370]: https://github.com/stackabletech/operator-rs/pull/370
 
-## [0.15.0] - 2022.03.21
+## [0.15.0] - 2022-03-21
 
 ### Added
 
@@ -28,7 +46,7 @@ All notable changes to this project will be documented in this file.
 [#348]: https://github.com/stackabletech/operator-rs/pull/348
 [#357]: https://github.com/stackabletech/operator-rs/pull/357
 
-## [0.14.1] - 2022.03.15
+## [0.14.1] - 2022-03-15
 
 ### Changed